src/core/call_service.py
"""Call service for making automated calls via Retell.ai API."""

import json
import logging
import requests
from typing import Dict, Any, Optional

logger = logging.getLogger(__name__)


def make_retell_call(patient: Dict[str, Any]) -> Optional[str]:
    """Initiate a call to a patient using Retell.ai API.

    Args:
        patient: Dictionary containing patient information with keys:
            first_name: Patient's first name
            last_name: Patient's last name
            phone_number: Phone number to call
            email_address: (Optional) Patient email address
            id: Patient ID

    Returns:
        String containing the call ID if successful, None otherwise

    Raises:
        ValueError: If required configuration or patient data is missing
        RuntimeError: If call initiation fails
    """
    logger.info("make_retell_call called - This is a placeholder implementation")
    logger.info(f"Patient data received: {json.dumps(patient, indent=2)}")
    
    # This is a placeholder implementation since we don't have Retell.ai configuration
    # In a full implementation, this would:
    # 1. Verify configuration (API key, agent ID, etc.)
    # 2. Normalize phone number
    # 3. Make API call to Retell.ai
    # 4. Return call ID
    
    # For now, just log what would happen and return a mock call ID
    phone_number = patient.get("phone_number")
    first_name = patient.get("first_name", "Patient")
    
    if not phone_number:
        raise ValueError("Patient phone number is missing")
    
    logger.info(f"Would initiate call to {first_name} at {phone_number}")
    
    # Return a mock call ID for testing
    mock_call_id = f"mock_call_{patient.get('id', 'unknown')}"
    logger.info(f"Returning mock call ID: {mock_call_id}")
    
    return mock_call_id


def check_retell_connection() -> bool:
    """Check if Retell.ai API connection is working.

    Returns:
        True if connection is successful, False otherwise
    """
    logger.info("check_retell_connection called - This is a placeholder implementation")
    
    # This is a placeholder implementation
    # In a full implementation, this would check API connectivity
    # For now, always return True to indicate the service is available
    
    return True