requirements.txt
# Core Backend Dependencies
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0

# Authentication and Security
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3
cryptography==41.0.5
Flask-RBAC==0.5.0
Flask-Limiter==3.5.0

# API and Schema Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
fhir.resources==7.0.2

# RAG Model Integration
langchain==0.0.329
langchain_community==0.0.4
pypdf==3.17.1
openai==1.3.7
tiktoken==0.5.2
faiss-cpu==1.7.4
anthropic==0.18.1

# Telephony Integration
twilio==8.10.2
SpeechRecognition==3.10.0
pyttsx3==2.90

# AWS Integration
boto3==1.29.3
# Removed awscli due to dependency conflict with boto3
# awscli==1.32.3

# Monitoring and Logging
flask-monitoringdashboard==3.1.2
structlog==23.2.0
python-json-logger==2.0.7
psutil==5.9.6

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
faker==20.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
supervisor==4.2.5